Remove red flag relating to use of articles and capitalization
In my experience procuring software from teams for whom English is a second language, errors with respect to articles, prepositions and capitalization are common. Such errors do not imply a lack of technical proficiency.

In any case, if a team lacks technical proficiency, then it will be clear from other parts of the proposal – a few stray articles, prepositions and letter cases won't be the sole evidence you need to rely on.

To avoid unfairly prejudicing the evaluators against teams for whom English is a second language, I suggest removing this list item.